requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
aiokafka==0.8.1
kafka-python==2.0.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic[email]==2.5.2
python-dotenv==1.0.0
langchain==0.0.339
langchain-openai==0.0.5
openai==1.3.5
qdrant-client==1.6.4
httpx==0.25.1
prometheus-client==0.19.0
pytest==7.4.3
requests==2.31.0
